PhishingDetectionBackend-main/requirements.txt
pandas
python-whois
beautifulsoup4
requests
xgboost
scikit-learn
pymongo